SF: Scheduler: set an expiration for heuristic layer's high refresh rate usages
Original change by Juhyung Park <qkrwngud825@gmail.com>,
refactored and adapted to Android 13.

Non-interactive heuristic layer's high refresh rate usages are overkill
and unnecessary.

Set a hard expiration limit to 3 seconds, or the idle timeout, whichever
is higher.

This effectively fixes/workarounds a major shortcoming in the whole
scheduler where it's totally unaware of the view's sizes.

For example, Instagram Reels with audios are played with a full-screen
video layer and a small animating audio icon at the bottom.

Because the audio icon's animations are lightweight, it's done at the
full, highest possible refresh rate. This prevents the device from
entering lower refresh rate states regardless of the video's fps.

With this commit, such layers are not allowed to use > 60 Hz states for
longer than 3 seconds.

[gotenksIN - Adapt to 14]

Change-Id: I065250bbbe3f0b9ff071779bcc43a4499463ef24
Co-authored-by: Juhyung Park <qkrwngud825@gmail.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Omkar Chandorkar <gotenksIN@aospa.co>
Signed-off-by: Kanishk <kanishkthederp@gmail.com>